big change
full convert to ejs, restructured the layout entirely, made use of partials

still many things that need to be done but functionally this works at the moment
