--- matcher-app/public/404 page/404.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>404: Page not Found</title>
-        <link rel="stylesheet" href="404Styles.css">
-    </head>
-    <body>
-        <div class="content">
-            <h1>404: Awwwwwww snap</h1>
-            <h3>Looks like the page you are looking for doesn't exist</h3>
-            <button href="#home">RETURN HOME</button>
-        </div>
-    </body>
-</html>
