+++ matcher-app/public/css/homepage.css
@@ -75,7 +75,7 @@ h1{
 }
 
 
-.card{
+.homepage-card{
     width: 100%;
     padding-left: 5%;
     position: relative;
