+++ matcher-app/public/css/navbar.css
@@ -1,0 +1,81 @@
+.navbar {
+    position: relative;
+}
+
+.brand {
+    position: absolute;
+    left: 50%;
+    margin-left: -50px !important;
+    margin-right: -50px !important;
+    display: block;
+    position: sticky;
+}
+@media screen and (min-width: 900px) {
+    .ph-only{
+        display:none;
+    }
+  }
+
+  @media screen and (max-width: 900px) {
+      .comp-only{
+          display:none;
+
+      }
+    }
+
+.img{
+    border-radius: 50%;
+}
+.navhome{
+    background-color: rgba(0,0,0,0.0) !important;
+}
+
+.homebut{
+    color: white !important;
+    font-weight: bold;
+    -webkit-transition: color 1s; /* For Safari 3.0 to 6.0 */
+     transition: color 1s; /* For modern browsers */
+}
+
+.homebut:hover {
+    filter: brightness(0);
+    color: #ff0000;
+}
+
+.roomtitle{
+    color: white;
+    font-weight: bold;
+    text-align-last: center;
+    font-size: 25px;
+}
+
+#bar .bar_item{
+    float: left;
+    padding-top: .5em;
+
+}
+
+.menu_item{
+    display:inline
+}
+
+#logo{
+    padding-right: 10em;
+    padding-left: 1em;
+    width: 100px;
+    font-weight: bold;
+
+}
+
+
+#bar_image{
+    display: flex;
+    justify-content: center;
+}
+
+.dropdown-menu-center {
+  left: 0% !important;
+  right: 50% !important;
+  text-align: center !important;
+  transform: translate(-55%, 0) !important;
+}
