--- matcher-app/public/home-page/homepage.html
@@ -1,63 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Starting Template</title>
-    <!-- Unlike the login and register page, we're using bootstrap 5 for everything else -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-iYQeCzEYFbKjA/T2uDLTpkwGzCiq6soy8tYaI1GyVh/UjpbCx/TYkiZhlZB6+fzT" crossorigin="anonymous">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.1/dist/js/bootstrap.bundle.min.js" integrity="sha384-u1OknCvxWvY5kfmNBILK2hRnQC3Pr17a+RTT6rIHI7NnikvbZlHgTPOOmMi466C8" crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="/homepage.css">
-</head>
-
-<body>
-    <nav class="navbar navbar-expand-lg navbar-dark bg-primary navhome">
-    <div class="container-fluid">
-        <a class="navbar-brand roomtitle" href="#">Roommate <br> Matcher</a>
-        <div class="rounded mx-auto d-block brand">
-            <a class="navbar-brand" href="#">
-                <img src="https://cdn.discordapp.com/attachments/1007427354685087784/1027709210533826691/unknown.png" width="60" height="48">
-            </a>
-        </div>
-        <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-            <span class="navbar-toggler-icon"></span>
-        </button>
-        <div class="collapse navbar-collapse justify-content-end" id="navbarNav">
-            <ul class="navbar-nav">
-                <li class="nav-item">
-                    <a class="nav-link active homebut" aria-current="page" href="#">Home</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link homebut" href="#">Dashboard</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link homebut" href="#">Login</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link homebut" href="#">Signup</a>
-                </li>
-            </ul>
-        </div>
-
-    </div>
-</nav>
-
-    <div class="card md bg-transparent mx-auto">
-        <div class="row g-0">
-          <div class=" col-xl-4">
-            <div class="card-body">
-              <p class="card-title">Need a Roomie?</p>
-              <p class ="card-text text-white"> Take our Quiz to match you with your perfect Roommate!</p>
-              <div class="input-group mb-2">
-                <button type="button" class="btn btn-primary btn-lg" >Find Your Roommate</button>
-              </div>
-            </div>
-          </div>
-          <div class=" col-xl-8 opacity-50">
-            <img src="../images/215-2154836_dorm-png-transparent-dorm-college-dorm-png.png" class="img-fluid w-100" alt="card-horizontal-image" id="card-image">
-          </div>
-        </div>
-      </div>
-
-</html>
