+++ matcher-app/public/start-template/start-template.ejs
@@ -1,0 +1,5 @@
+<%- include('../partials/header'); %>
+    <h1>Hello World!</h1>
+</body>
+
+</html>
