--- matcher-app/public/styles.css
@@ -1,112 +1,0 @@
-
-html{
-    height: 100vh;
-    overflow-y: hidden;
-    
-}
-body {
-    background: linear-gradient(116.82deg, #1141EC 0%, #006EFF 43.82%, #009FFF 100%);
-    overflow-y: scroll;
-    height: 100vh;
-}
-
-h1{
-    color: white;
-}
-
-
-.navbar {
-    position: relative;
-}
-
-.brand {
-    position: absolute;
-    left: 50%;
-    margin-left: -50px !important;
-    margin-right: -50px !important;
-    display: block;
-    position: sticky;
-}
-
-
-.navhome{
-    background-color: rgba(0,0,0,0.0) !important;
-}
-
-.homebut{
-    color: white !important;
-    font-weight: bold;
-    -webkit-transition: color 1s; /* For Safari 3.0 to 6.0 */
-     transition: color 1s; /* For modern browsers */
-}
-
-.homebut:hover {
-    filter: brightness(0);
-    color: #ff0000;
-}
-.roomtitle{
-    color: white;
-    font-weight: bold;
-    text-align-last: center;
-    font-size: 25px;
-}
-
-#bar .bar_item{
-    float: left;
-    padding-top: .5em;
-
-}
-
-.menu_item{
-    display:inline
-}
-
-#logo{
-    padding-right: 10em;
-    padding-left: 1em;
-    width: 100px;
-    font-weight: bold;
-
-}
-
-
-#bar_image{
-    display: flex;
-    justify-content: center;
-}
-
-
-.card{
-    width: 100%;
-    padding-left: 5%;
-    position: relative;
-    top: 15%;
-    border: none !important; 
-    border-radius: 1em;
-}
-
-
-.card-title{
-    color: white;
-    font-weight: bold;
-    text-align-last: left;
-    font-size: 40px;
-}
-
-/*This is modifying the btn-primary colors but you could create your own .btn-something class as well*/
-.btn-primary {
-    color: black;
-    background-color: white;
-    border-color: gray; /*set the color you want here*/
-}
-.btn-primary:hover, .btn-primary:focus, .btn-primary:active, .btn-primary.active, .open>.dropdown-toggle.btn-primary {
-    color: #fff;
-    background-color: rgb(0, 38, 255) !important; 
-    border-color: gray; /*set the color you want here*/
-}
-
-
-
-
-
-
