+++ matcher-app/views/404/404.ejs
@@ -1,0 +1,8 @@
+<%- include('../partials/header'); %>
+        <div class="content">
+            <h1>404: Awwwwwww snap</h1>
+            <h3>Looks like the page you are looking for doesn't exist.</h3>
+            <a href="/"><button>RETURN HOME</button></a>
+        </div>
+    </body>
+</html>
