+++ matcher-app/views/dashboard/dashboard.ejs
@@ -1,15 +1,5 @@
 <!--Author: Preston Button https://github.com/prestonButton -->
-
-<!DOCTYPE html>
-<html lang="en-US">
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>Matches Box</title>
-        <link rel="stylesheet" href="matchesBox.css">
-    </head>
-    <body>
-        <!--Add Navbar element here when it has been completed-->    
+<%- include('../partials/header'); %>
         <div class="div1">
             <h2>Your Matches!</h2>
             <div class="matchBox">
