+++ matcher-app/views/homepage/homepage.ejs
@@ -1,0 +1,19 @@
+<%- include('../partials/header'); %>
+    <div class="homepage-card md bg-transparent mx-auto">
+        <div class="row g-0">
+          <div class=" col-xl-4">
+            <div class="card-body">
+              <p class="card-title">Need a Roomie?</p>
+              <p class ="card-text text-white"> Take our Quiz to match you with your perfect Roommate!</p>
+              <div class="input-group mb-2">
+                <button type="button" class="btn btn-primary btn-lg" >Find Your Roommate</button>
+              </div>
+            </div>
+          </div>
+          <div class=" col-xl-8 opacity-50">
+            <img src="../images/215-2154836_dorm-png-transparent-dorm-college-dorm-png.png" class="img-fluid w-100" alt="card-horizontal-image" id="card-image">
+          </div>
+        </div>
+      </div>
+
+</html>
