+++ matcher-app/views/partials/header.ejs
@@ -1,0 +1,60 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Roommate Matcher</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-iYQeCzEYFbKjA/T2uDLTpkwGzCiq6soy8tYaI1GyVh/UjpbCx/TYkiZhlZB6+fzT" crossorigin="anonymous">
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.1/dist/js/bootstrap.bundle.min.js" integrity="sha384-u1OknCvxWvY5kfmNBILK2hRnQC3Pr17a+RTT6rIHI7NnikvbZlHgTPOOmMi466C8" crossorigin="anonymous"></script>
+    <link rel="stylesheet" href="css/<%=styleInput%>.css">
+    <link rel="stylesheet" href="css/homepage.css">
+    <link rel="stylesheet" href="css/navbar.css">
+</head>
+<body>
+    <nav class="navbar navbar-expand-lg navbar-dark bg-primary navhome">
+        <div class="container-fluid">
+            <a class="navbar-brand roomtitle" href="/">Roommate <br> Matcher</a>
+            <div class="rounded mx-auto d-block brand">
+                <a class="navbar-brand" href="">
+                    <img src="https://cdn.discordapp.com/attachments/1007427354685087784/1027709210533826691/unknown.png" width="60" height="48">
+                </a>
+            </div>
+            <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
+                <span class="navbar-toggler-icon"></span>
+            </button>
+            <div class="collapse navbar-collapse justify-content-end" id="navbarNav">
+                <ul class="navbar-nav">
+                    <li class="nav-item">
+                        <a class="nav-link active homebut" aria-current="page" href="/">Home</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link homebut" href="/dashboard">Dashboard</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="ph-only nav-link homebut" href="">Your Profile</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="ph-only nav-link homebut" href="">Change an Answer</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="ph-only nav-link homebut" href="">Report a Problem</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="ph-only nav-link homebut" href="">Signout</a>
+                    </li>
+                    <li class="nav-item dropdown">
+                        <a class="nav-link dropdown-toggle comp-only" href="#" id="navbarDropdown" role="button" data-bs-toggle="dropdown" aria-expanded="true">
+                            <img class="img" src="https://cdn.discordapp.com/attachments/1007427354685087784/1035254854802743307/UserIcon.png" width="30" height="24">
+                        </a>
+                        <ul class="dropdown-menu dropdown-menu-center" aria-labelledby="navbarDropdown">
+                            <li><a class="dropdown-item" href="">Your Profile</a></li>
+                            <li><a class="dropdown-item" href="">Change an Answer</a></li>
+                            <li><a class="dropdown-item" href="">Report a Problem</a></li>
+                            <li><a class="dropdown-item" href="">Signout</a></li>
+                        </ul>
+                    </li>
+                </ul>
+            </div>
+        </div>
+    </nav>
