+++ matcher-app/views/problem/problem.ejs
@@ -1,12 +1,4 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Report A Problem</title>
-    <link rel="stylesheet" href="problem.css" />
-  </head>
-  <body>
+<%- include('../partials/header'); %>
     <h1>Report A Problem</h1>
     <form>
     <fieldset>
