+++ matcher-app/views/register/register.ejs
@@ -1,16 +1,4 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-
-<head>
-  <meta charset="utf-8">
-  <title>Secrets</title>
-  <!-- Keep in mind we're using bootstrap 4.2.1 for this page, if you're trying to debug a bootstrap issue, check the docs for bootstrap 4, not 5 -->
-  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.6.3/css/all.css" integrity="sha384-UHRtZLI+pbxtHCWp1t77Bi1L4ZtiqrqD80Kn4Z8NTSRyMA2Fd33n5dQ8lWUE00s/" crossorigin="anonymous">
-  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.2.1/css/bootstrap.min.css" integrity="sha384-GJzZqFGwb1QTTN6wy59ffF1BuGJpLSa9DkKMp0DgiMDm4iYMj70gZWKYbI706tWS" crossorigin="anonymous">
-  <link rel="stylesheet" href="css/styles.css">
-</head>
-
-<body>
+<%- include('../partials/header'); %>
 <div class="container mt-5">
   <h1>Register</h1>
   <div class="row">
@@ -29,7 +17,6 @@
                 <label for="lastName">Last Name</label>
                 <input type="text" class="form-control" name="lastName">
               </div>
-                <input type="text" class="form-control" name="lastName">           </div>
             </div>
             <div style="display: flex;">
               <div class="form-group" style="width: 45%; ">
@@ -69,15 +56,9 @@
             </div>
             <button type="submit" class="btn btn-dark">Register</button>
           </form>
-
         </div>
       </div>
     </div>
   </div>
 </div>
-
-
-
-</body>
-</html>
